+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/questionableActivity/ListingQuestionableActivityProvider.java
@@ -1,78 +1,241 @@
 package gov.healthit.chpl.questionableActivity;
 
+import java.time.ZoneOffset;
 import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.Date;
 import java.util.List;
+import java.util.TimeZone;
 
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 
 import gov.healthit.chpl.domain.CQMResultDetails;
 import gov.healthit.chpl.domain.CertificationResult;
+import gov.healthit.chpl.domain.CertificationStatusEvent;
 import gov.healthit.chpl.domain.CertifiedProductSearchDetails;
 import gov.healthit.chpl.dto.questionableActivity.QuestionableActivityListingDTO;
 import gov.healthit.chpl.entity.CertificationStatusType;
 
+/**
+ * Parses activity on Listings to see if there are questionable activities.
+ * @author alarned
+ *
+ */
 @Component
 public class ListingQuestionableActivityProvider {
-    
+
+    /**
+     * Create questionable activity if the listing was a 2011 listing.
+     * @param origListing original listing
+     * @param newListing new listing
+     * @return questionable activity, if it exists
+     */
     public QuestionableActivityListingDTO check2011EditionUpdated(
-            CertifiedProductSearchDetails origListing, CertifiedProductSearchDetails newListing) {
-        
+            final CertifiedProductSearchDetails origListing, final CertifiedProductSearchDetails newListing) {
+
         QuestionableActivityListingDTO activity = null;
         if (origListing.getCertificationEdition().get("name").equals("2011")) {
               activity = new QuestionableActivityListingDTO();
               activity.setBefore(null);
               activity.setAfter(null);
         }
-        
+
         return activity;
     }
-    
-    public QuestionableActivityListingDTO checkCertificationStatusUpdated(
-            CertifiedProductSearchDetails origListing, CertifiedProductSearchDetails newListing) {
-        
-        QuestionableActivityListingDTO activity = null;
-        if (!origListing.getCurrentStatus().getStatus().getId()
-                .equals(newListing.getCurrentStatus().getStatus().getId())) {
-              activity = new QuestionableActivityListingDTO();
-              activity.setBefore(origListing.getCurrentStatus().getStatus().getName());
-              activity.setAfter(newListing.getCurrentStatus().getStatus().getName());
-              activity.setCertificationStatusChangeReason(newListing.getCurrentStatus().getReason());
-        }
-        
-        return activity;
-    }
-    
+
     /**
-     * questionable only if the certification status has updated
-     * to the supplied updateTo value
-     * @param updateTo
-     * @param origListing
-     * @param newListing
-     * @return
+     * Create questionable activity if the current certification status was updated.
+     * @param origListing original listing
+     * @param newListing new listing
+     * @return questionable activity if it exists
      */
     public QuestionableActivityListingDTO checkCertificationStatusUpdated(
-            CertificationStatusType updateTo, CertifiedProductSearchDetails origListing, 
-            CertifiedProductSearchDetails newListing) {
-        
+            final CertifiedProductSearchDetails origListing, final CertifiedProductSearchDetails newListing) {
+
         QuestionableActivityListingDTO activity = null;
-        if (!origListing.getCurrentStatus().getStatus().getName().equals(updateTo.getName()) && 
-            newListing.getCurrentStatus().getStatus().getName().equals(updateTo.getName())) {
+        CertificationStatusEvent prev = origListing.getCurrentStatus();
+        CertificationStatusEvent curr = newListing.getCurrentStatus();
+        if (!prev.getStatus().getId().equals(curr.getStatus().getId())) {
+              activity = new QuestionableActivityListingDTO();
+              activity.setBefore(prev.getStatus().getName());
+              activity.setAfter(curr.getStatus().getName());
+              activity.setCertificationStatusChangeReason(curr.getReason());
+        }
+
+        return activity;
+    }
+
+    /**
+     * Create questionable activity if the historical certification statuses or dates were updated.
+     * Sorts certification status events by date, earliest first, then compares them, incrementing
+     * through them with the "earlier" date getting moved up. If both are equal both move.
+     * Does not compare "latest" events, as those are the "current" values, and compared in other functions.
+     * @param origListing original listing
+     * @param newListing new listing
+     * @return questionable activity if it exists
+     */
+    public QuestionableActivityListingDTO checkCertificationStatusHistoryUpdated(
+            final CertifiedProductSearchDetails origListing, final CertifiedProductSearchDetails newListing) {
+
+        QuestionableActivityListingDTO activity = null;
+        List<CertificationStatusEvent> prevEvents = origListing.getCertificationEvents();
+        List<CertificationStatusEvent> currEvents = newListing.getCertificationEvents();
+        Collections.sort(prevEvents, new CertificationStatusEventComparator());
+        Collections.sort(currEvents, new CertificationStatusEventComparator());
+        int p = 0, c = 0;
+        List<String> beforeRes = new ArrayList<String>();
+        List<String> afterRes = new ArrayList<String>();
+        List<String> reasRes = new ArrayList<String>();
+        String pRes, cRes, reas;
+        CertificationStatusEvent pEvent, cEvent;
+        Calendar displayDate = Calendar.getInstance(TimeZone.getTimeZone(ZoneOffset.UTC));
+        while (p < prevEvents.size() - 1 && c < currEvents.size() - 1) {
+            pRes = "";
+            cRes = "";
+            reas = "";
+            pEvent = prevEvents.get(p);
+            cEvent = currEvents.get(c);
+            if (pEvent.getEventDate().longValue() < cEvent.getEventDate().longValue()) {
+                if (pEvent.getStatus().getId().equals(cEvent.getStatus().getId())) {
+                    displayDate.setTimeInMillis(pEvent.getEventDate());
+                    pRes = pEvent.getStatus().getName() + " (" + displayDate.getTime().toString() + ")";
+                    displayDate.setTimeInMillis(cEvent.getEventDate());
+                    cRes = cEvent.getStatus().getName() + " (" + displayDate.getTime().toString() + ")";
+                    reas = cEvent.getReason();
+                    p += 1;
+                    c += 1;
+                } else {
+                    displayDate.setTimeInMillis(pEvent.getEventDate());
+                    pRes = pEvent.getStatus().getName() + " (" + displayDate.getTime().toString() + ")";
+                    cRes = "Removed";
+                    reas = pEvent.getReason();
+                    p += 1;
+                }
+            } else if (pEvent.getEventDate().longValue() > cEvent.getEventDate().longValue()) {
+                if (pEvent.getStatus().getId().equals(cEvent.getStatus().getId())) {
+                    displayDate.setTimeInMillis(pEvent.getEventDate());
+                    pRes = pEvent.getStatus().getName() + " (" + displayDate.getTime().toString() + ")";
+                    displayDate.setTimeInMillis(cEvent.getEventDate());
+                    cRes = cEvent.getStatus().getName() + " (" + displayDate.getTime().toString() + ")";
+                    reas = cEvent.getReason();
+                    p += 1;
+                    c += 1;
+                } else {
+                    pRes = "Added";
+                    displayDate.setTimeInMillis(cEvent.getEventDate());
+                    cRes = cEvent.getStatus().getName() + " (" + displayDate.getTime().toString() + ")";
+                    reas = cEvent.getReason();
+                    c += 1;
+                }
+            } else if (!pEvent.getStatus().getId().equals(cEvent.getStatus().getId())) {
+                displayDate.setTimeInMillis(pEvent.getEventDate());
+                pRes = pEvent.getStatus().getName() + " (" + displayDate.getTime().toString() + ")";
+                displayDate.setTimeInMillis(cEvent.getEventDate());
+                cRes = cEvent.getStatus().getName() + " (" + displayDate.getTime().toString() + ")";
+                reas = cEvent.getReason();
+                p += 1;
+                c += 1;
+            } else {
+                p += 1;
+                c += 1;
+            }
+            if (pRes.length() > 0 || cRes.length() > 0) {
+                beforeRes.add(pRes);
+                afterRes.add(cRes);
+                reasRes.add(reas);
+            }
+        }
+        while (p < prevEvents.size() - 1) {
+            pEvent = prevEvents.get(p);
+            displayDate.setTimeInMillis(pEvent.getEventDate());
+            beforeRes.add(pEvent.getStatus().getName() + " (" + displayDate.getTime().toString() + ")");
+            afterRes.add("Removed");
+            reasRes.add(pEvent.getReason());
+            p += 1;
+        }
+        while (c < currEvents.size() - 1) {
+            cEvent = currEvents.get(c);
+            beforeRes.add("Added");
+            displayDate.setTimeInMillis(cEvent.getEventDate());
+            afterRes.add(cEvent.getStatus().getName() + " (" + displayDate.getTime().toString() + ")");
+            reasRes.add(cEvent.getReason());
+            c += 1;
+        }
+        if (beforeRes.size() > 0 || afterRes.size() > 0) {
+            activity = new QuestionableActivityListingDTO();
+            activity.setBefore(beforeRes.toString());
+            activity.setAfter(afterRes.toString());
+            activity.setCertificationStatusChangeReason(reasRes.toString());
+        }
+
+        return activity;
+    }
+
+
+   /**
+     * Create questionable activity if the current certification status event date was updated.
+     * @param origListing original listing
+     * @param newListing new listing
+     * @return questionable activity if it exists
+     */
+    public QuestionableActivityListingDTO checkCertificationStatusDateUpdated(
+            final CertifiedProductSearchDetails origListing, final CertifiedProductSearchDetails newListing) {
+
+        QuestionableActivityListingDTO activity = null;
+        CertificationStatusEvent prev = origListing.getCurrentStatus();
+        CertificationStatusEvent curr = newListing.getCurrentStatus();
+        Calendar displayDate = Calendar.getInstance(TimeZone.getTimeZone(ZoneOffset.UTC));
+        if (prev.getEventDate().longValue() != curr.getEventDate().longValue()) {
+              activity = new QuestionableActivityListingDTO();
+              displayDate.setTimeInMillis(prev.getEventDate());
+              activity.setBefore(displayDate.getTime().toString());
+              displayDate.setTimeInMillis(curr.getEventDate());
+              activity.setAfter(displayDate.getTime().toString());
+              activity.setCertificationStatusChangeReason(curr.getReason());
+        }
+
+        return activity;
+    }
+
+    /**
+     * questionable only if the certification status has updated
+     * to the supplied updateTo value.
+     * @param updateTo status to check against
+     * @param origListing original listing
+     * @param newListing new listing
+     * @return activity if it is questionable
+     */
+    public QuestionableActivityListingDTO checkCertificationStatusUpdated(
+            final CertificationStatusType updateTo, final CertifiedProductSearchDetails origListing,
+            final CertifiedProductSearchDetails newListing) {
+
+        QuestionableActivityListingDTO activity = null;
+        if (!origListing.getCurrentStatus().getStatus().getName().equals(updateTo.getName())
+                && newListing.getCurrentStatus().getStatus().getName().equals(updateTo.getName())) {
               activity = new QuestionableActivityListingDTO();
               activity.setBefore(origListing.getCurrentStatus().getStatus().getName());
               activity.setAfter(newListing.getCurrentStatus().getStatus().getName());
               activity.setCertificationStatusChangeReason(newListing.getCurrentStatus().getReason());
         }
-        
+
         return activity;
     }
-    
+
+    /**
+     * Create questionable activity if CQMs were added.
+     * @param origListing original listing
+     * @param newListing new listing
+     * @return questionable activity, if it exists
+     */
     public List<QuestionableActivityListingDTO> checkCqmsAdded(
-            CertifiedProductSearchDetails origListing, CertifiedProductSearchDetails newListing) {
-        
+            final CertifiedProductSearchDetails origListing, final CertifiedProductSearchDetails newListing) {
+
         List<QuestionableActivityListingDTO> cqmAddedActivities = new ArrayList<QuestionableActivityListingDTO>();
-        if (origListing.getCqmResults() != null && origListing.getCqmResults().size() > 0 && 
-                newListing.getCqmResults() != null && newListing.getCqmResults().size() > 0) {
+        if (origListing.getCqmResults() != null && origListing.getCqmResults().size() > 0
+                && newListing.getCqmResults() != null && newListing.getCqmResults().size() > 0) {
             //all cqms are in the details so find the same one in the orig and new objects
             //based on cms id and compare the success boolean to see if one was added
             for (CQMResultDetails origCqm : origListing.getCqmResults()) {
@@ -107,16 +270,21 @@ public class ListingQuestionableActivityProvider {
                 }
             }
         }
-        
         return cqmAddedActivities;
     }
-    
+
+    /**
+     * Create questionable activity if CQMs were removed.
+     * @param origListing original listing
+     * @param newListing new listing
+     * @return questionable activity, if it exists
+     */
     public List<QuestionableActivityListingDTO> checkCqmsRemoved(
-            CertifiedProductSearchDetails origListing, CertifiedProductSearchDetails newListing) {
-        
+            final CertifiedProductSearchDetails origListing, final CertifiedProductSearchDetails newListing) {
+
         List<QuestionableActivityListingDTO> cqmRemovedActivities = new ArrayList<QuestionableActivityListingDTO>();
-        if (origListing.getCqmResults() != null && origListing.getCqmResults().size() > 0 && 
-                newListing.getCqmResults() != null && newListing.getCqmResults().size() > 0) {
+        if (origListing.getCqmResults() != null && origListing.getCqmResults().size() > 0
+                && newListing.getCqmResults() != null && newListing.getCqmResults().size() > 0) {
             //all cqms are in the details so find the same one in the orig and new objects
             //based on cms id and compare the success boolean to see if one was removed
             for (CQMResultDetails origCqm : origListing.getCqmResults()) {
@@ -131,7 +299,8 @@ public class ListingQuestionableActivityProvider {
                         if (origCqm.isSuccess() == Boolean.TRUE && newCqm.isSuccess() == Boolean.FALSE) {
                             //orig did have this cqm but new does not so it was removed
                             QuestionableActivityListingDTO activity = new QuestionableActivityListingDTO();
-                            activity.setBefore(origCqm.getCmsId() != null ? origCqm.getCmsId() : origCqm.getNqfNumber());
+                            activity.setBefore(
+                                    origCqm.getCmsId() != null ? origCqm.getCmsId() : origCqm.getNqfNumber());
                             activity.setAfter(null);
                             cqmRemovedActivities.add(activity);
                         }
@@ -142,13 +311,15 @@ public class ListingQuestionableActivityProvider {
                         if (origCqm.isSuccess() == Boolean.TRUE && newCqm.isSuccess() == Boolean.FALSE) {
                             //orig did not have this cqm but new does so it was added
                             QuestionableActivityListingDTO activity = new QuestionableActivityListingDTO();
-                            activity.setBefore(origCqm.getCmsId() != null ? origCqm.getCmsId() : origCqm.getNqfNumber());
+                            activity.setBefore(
+                                    origCqm.getCmsId() != null ? origCqm.getCmsId() : origCqm.getNqfNumber());
                             activity.setAfter(null);
                             cqmRemovedActivities.add(activity);
-                        } else if(origCqm.getSuccessVersions() != null && origCqm.getSuccessVersions().size() > 0 && 
-                                (newCqm.getSuccessVersions() == null || newCqm.getSuccessVersions().size() == 0)) {
+                        } else if (origCqm.getSuccessVersions() != null && origCqm.getSuccessVersions().size() > 0
+                                && (newCqm.getSuccessVersions() == null || newCqm.getSuccessVersions().size() == 0)) {
                             QuestionableActivityListingDTO activity = new QuestionableActivityListingDTO();
-                            activity.setBefore(origCqm.getCmsId() != null ? origCqm.getCmsId() : origCqm.getNqfNumber());
+                            activity.setBefore(
+                                    origCqm.getCmsId() != null ? origCqm.getCmsId() : origCqm.getNqfNumber());
                             activity.setAfter(null);
                             cqmRemovedActivities.add(activity);
                         }
@@ -157,16 +328,21 @@ public class ListingQuestionableActivityProvider {
                 }
             }
         }
-        
         return cqmRemovedActivities;
     }
-    
+
+    /**
+     * Create questionable activity if certification criteria were added.
+     * @param origListing original listing
+     * @param newListing new listing
+     * @return questionable activity, if it exists
+     */
     public List<QuestionableActivityListingDTO> checkCertificationsAdded(
-            CertifiedProductSearchDetails origListing, CertifiedProductSearchDetails newListing) {
-        
+            final CertifiedProductSearchDetails origListing, final CertifiedProductSearchDetails newListing) {
+
         List<QuestionableActivityListingDTO> certAddedActivities = new ArrayList<QuestionableActivityListingDTO>();
-        if (origListing.getCertificationResults() != null && origListing.getCertificationResults().size() > 0 && 
-                newListing.getCertificationResults() != null && newListing.getCertificationResults().size() > 0) {
+        if (origListing.getCertificationResults() != null && origListing.getCertificationResults().size() > 0
+                && newListing.getCertificationResults() != null && newListing.getCertificationResults().size() > 0) {
             //all cert results are in the details so find the same one in the orig and new objects
             //based on number and compare the success boolean to see if one was added
             for (CertificationResult origCertResult : origListing.getCertificationResults()) {
@@ -187,13 +363,19 @@ public class ListingQuestionableActivityProvider {
 
         return certAddedActivities;
     }
-    
+
+    /**
+     * Create questionable activity if it has removal of certification criteria.
+     * @param origListing original listing
+     * @param newListing new listing
+     * @return questionable activity, if it is
+     */
     public List<QuestionableActivityListingDTO> checkCertificationsRemoved(
-            CertifiedProductSearchDetails origListing, CertifiedProductSearchDetails newListing) {
-        
+            final CertifiedProductSearchDetails origListing, final CertifiedProductSearchDetails newListing) {
+
         List<QuestionableActivityListingDTO> certRemovedActivities = new ArrayList<QuestionableActivityListingDTO>();
-        if (origListing.getCertificationResults() != null && origListing.getCertificationResults().size() > 0 && 
-                newListing.getCertificationResults() != null && newListing.getCertificationResults().size() > 0) {
+        if (origListing.getCertificationResults() != null && origListing.getCertificationResults().size() > 0
+                && newListing.getCertificationResults() != null && newListing.getCertificationResults().size() > 0) {
             //all cert results are in the details so find the same one in the orig and new objects
             //based on number and compare the success boolean to see if one was removed
             for (CertificationResult origCertResult : origListing.getCertificationResults()) {
@@ -211,23 +393,35 @@ public class ListingQuestionableActivityProvider {
                 }
             }
         }
-        
         return certRemovedActivities;
     }
-    
+
+    /**
+     * Check to see if activity has has deletion of surveillance.
+     * @param origListing original listing
+     * @param newListing new listing
+     * @return activity if it is questionable
+     */
     public QuestionableActivityListingDTO checkSurveillanceDeleted(
-            CertifiedProductSearchDetails origListing, CertifiedProductSearchDetails newListing) {
-        
+            final CertifiedProductSearchDetails origListing, final CertifiedProductSearchDetails newListing) {
+
         QuestionableActivityListingDTO activity = null;
-        if(origListing.getSurveillance() != null && origListing.getSurveillance().size() > 0 && 
-                (newListing.getSurveillance() == null || 
-                 newListing.getSurveillance().size() < origListing.getSurveillance().size())) {
-              
+        if (origListing.getSurveillance() != null && origListing.getSurveillance().size() > 0
+                && (newListing.getSurveillance() == null
+                || newListing.getSurveillance().size() < origListing.getSurveillance().size())) {
+
               activity = new QuestionableActivityListingDTO();
               activity.setBefore(null);
               activity.setAfter(null);
         }
-        
         return activity;
+    }
+
+    class CertificationStatusEventComparator implements Comparator<CertificationStatusEvent> {
+        @Override
+        public int compare(final CertificationStatusEvent a, final CertificationStatusEvent b) {
+            return a.getEventDate().longValue() < b.getEventDate().longValue()
+                    ? -1 : a.getEventDate().longValue() == b.getEventDate().longValue() ? 0 : 1;
+        }
     }
 }